% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contacts.R
\docType{methods}
\name{addFeature}
\alias{addFeature}
\alias{addFeature,contacts,GRangesOrGInteractions,character-method}
\alias{addFeature,contacts,GRangesOrGInteractions-method}
\title{addFeature}
\usage{
addFeature(x, feature, feature.name)

\S4method{addFeature}{contacts,GRangesOrGInteractions,character}(x, feature, feature.name)
}
\arguments{
\item{x}{a contacts object}

\item{feature}{a GRanges or GInteractions object}

\item{feature.name}{a name to give to the new features}
}
\value{
The input contacts object, with a newly added feature
}
\description{
addFeature
}
