% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_track.R
\name{addTracks}
\alias{addTracks}
\title{addTracks}
\usage{
addTracks(p, range, annotations = NULL, profiles = NULL)
}
\arguments{
\item{p}{p}

\item{range}{range}

\item{annotations}{annotations}

\item{profiles}{profiles}
}
\description{
addTracks
}
