% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{AllGenerics}
\alias{AllGenerics}
\alias{plotMatrix}
\title{Generic functions}
\arguments{
\item{x}{Can be a \code{HiCExperiment}, a GInteractions or an array/matrix object.}

\item{use.scores}{use.scores}

\item{scale}{scale}

\item{limits}{limits}

\item{max.distance}{maximum distance. If provided, the heatmap is plotted
horizontally.}

\item{loops}{loops}

\item{borders}{borders}

\item{dpi}{dpi}

\item{rasterize}{rasterize}

\item{symmetrical}{symmetrical}

\item{chrom_lines}{chrom_lines}

\item{show_grid}{show_grid}

\item{cmap}{color map}
}
\description{
Generics functions created in HiContacts package.
}
