% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APA.R
\name{APA}
\alias{APA}
\title{APA}
\usage{
APA(x, coords, bins = 50, use.scores = "balanced")
}
\arguments{
\item{x}{x}

\item{coords}{coords}

\item{bins}{bins}

\item{use.scores}{use.scores}
}
\value{
an aggregated contact matrix
}
\description{
APA
}
