% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cistrans.R
\name{cis_trans}
\alias{cis_trans}
\title{cis_trans}
\usage{
cis_trans(x)
}
\arguments{
\item{x}{A \code{contacts} object over the full genome}
}
\value{
a tibble, listing for each chr. the \% of cis/trans interactions
}
\description{
cis_trans
}
\examples{
library(HiContacts)
data(full_contacts_yeast)
cis_trans(full_contacts_yeast)
}
