% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cistrans.R
\name{cisTransRatio}
\alias{cisTransRatio}
\title{cisTransRatio}
\usage{
cisTransRatio(x)
}
\arguments{
\item{x}{A \code{HiCExperiment} object over the full genome}
}
\value{
a tibble, listing for each chr. the \% of cis/trans interactions
}
\description{
Quickly computes a cis-trans ratio of interactions.
}
\examples{
library(HiContacts)
full_contacts_yeast <- contacts_yeast(full = TRUE)
cisTransRatio(full_contacts_yeast)
}
