% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartments.R
\name{compartments}
\alias{compartments}
\alias{getCompartments}
\title{Contact map compartments}
\usage{
getCompartments(
  x,
  resolution = NULL,
  genome = NULL,
  chromosomes = NULL,
  neigens = 3,
  autocorrelation = TRUE,
  BPPARAM = BiocParallel::bpparam()
)
}
\arguments{
\item{x}{A \code{HiCExperiment} object over a full genome}

\item{resolution}{Which resolution to use to compute eigen vectors}

\item{genome}{a BSgenome of DNAStringSet object associated with the Hi-C
contact matrix.}

\item{chromosomes}{character or integer vector indicating which}

\item{neigens}{Numver of eigen vectors to extract}

\item{autocorrelation}{Whether to perform matrix autocorrelation prior to
computing the eigenvectors}

\item{BPPARAM}{BiocParallel parallelization settings}
}
\value{
A \code{HiCExperiment} object with additional \code{eigens} metadata containing the
normalized eigenvectors and a new "compartments" topologicalFeatures
storing A and B compartments as a GRanges object.
}
\description{
Computes eigen vectors for each chromosome using cis contacts and extract
chromosome compartments.
}
\examples{
library(HiContacts)
full_contacts_yeast <- full_contacts_yeast()
comps <- getCompartments(full_contacts_yeast)
S4Vectors::metadata(comps)$eigens
}
