% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contacts.R
\docType{class}
\name{contacts-class}
\alias{contacts-class}
\title{contacts S4 class}
\description{
An S4 class to represent a file-stored (as .(m)cool) contact matrix
imported in R.
}
\section{Slots}{

\describe{
\item{\code{focus}}{Chr. coordinates for which interaction counts are extracted
from the .(m)cool file.}

\item{\code{metadata}}{<etadata associated with the .(m)cool file.}

\item{\code{seqinfo}}{Seqinfo, deduced from the largest resolution available
in the .(m)cool file.}

\item{\code{resolutions}}{Resolutions available in the .(m)cool file.}

\item{\code{current_resolution}}{Current resolution}

\item{\code{bins}}{Genomic bins (computed on-the-fly) of seqinfo at the current
resolution}

\item{\code{interactions}}{Genomic Interactions extracted from the .(m)cool object}

\item{\code{assays}}{Available interaction scores.}

\item{\code{features}}{Genomic features associated with the dataset (e.g.
loops, borders, etc...)}

\item{\code{pairsFile}}{path for the .pairs file associated with the .(m)cool file}

\item{\code{type}}{Optional. Type of contacts matrix (sparse, full, aggr, ratio, ...)}
}}

