% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contacts.R
\docType{class}
\name{GRangesOrGInteractions-class}
\alias{GRangesOrGInteractions-class}
\alias{contacts-class}
\alias{contacts}
\alias{length}
\alias{length,contacts-method}
\alias{dim}
\alias{dim,contacts-method}
\alias{[}
\alias{[,contacts-method}
\alias{show}
\alias{show,contacts-method}
\title{contacts S4 class}
\usage{
contacts(
  cool_path,
  resolution = NULL,
  focus = NULL,
  metadata = NULL,
  features = NULL,
  pairs = NULL
)

\S4method{length}{contacts}(x)

\S4method{dim}{contacts}(x)

\S4method{[}{contacts,ANY,ANY,ANY}(x, i)

\S4method{show}{contacts}(object)
}
\arguments{
\item{x}{A \code{contacts} object.}
}
\description{
An S4 class to represent a file-stored (as .(m)cool) contact matrix
imported in R.
}
\section{Slots}{

\describe{
\item{\code{focus}}{Chr. coordinates for which interaction counts are extracted
from the .(m)cool file.}

\item{\code{metadata}}{<etadata associated with the .(m)cool file.}

\item{\code{seqinfo}}{Seqinfo, deduced from the largest resolution available
in the .(m)cool file.}

\item{\code{resolutions}}{Resolutions available in the .(m)cool file.}

\item{\code{current_resolution}}{Current resolution}

\item{\code{bins}}{Genomic bins (computed on-the-fly) of seqinfo at the current
resolution}

\item{\code{interactions}}{Genomic Interactions extracted from the .(m)cool object}

\item{\code{assays}}{Available interaction scores.}

\item{\code{features}}{Genomic features associated with the dataset (e.g.
loops, borders, etc...)}

\item{\code{pairsFile}}{path for the .pairs file associated with the .(m)cool file}

\item{\code{type}}{Optional. Type of contacts matrix (sparse, full, aggr, ratio, ...)}
}}

