% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{contacts_yeast}
\alias{contacts_yeast}
\title{Example of a \code{contacts} object}
\format{
An object of class \code{contacts} with 802 rows and 802 columns.
}
\source{
HiContactsData

fpath <- HiContactsData::HiContactsData('yeast_wt', 'mcool')
contacts_yeast <- contacts(fpath, 'II', resolution = 1000)
}
\usage{
contacts_yeast
}
\description{
Example of a \code{contacts} object
}
\keyword{datasets}
