% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{contacts_yeast_eco1}
\alias{contacts_yeast_eco1}
\title{Example of a \code{contacts} object}
\format{
An object of class \code{contacts} with 802 rows and 802 columns.
}
\source{
HiContactsData

fpath <- HiContactsData::HiContactsData('yeast_eco1', 'mcool')
contacts_yeast_eco1 <- contacts(fpath, 'II', resolution = 1000)
}
\usage{
contacts_yeast_eco1
}
\description{
Example of a \code{contacts} object
}
\keyword{datasets}
