% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{contacts_yeast}
\alias{contacts_yeast}
\alias{contacts_yeast_eco1}
\alias{full_contacts_yeast}
\title{Example datasets provided in \code{HiContacts} & \code{HiContactsData}}
\source{
HiContacts
}
\usage{
contacts_yeast()

contacts_yeast_eco1()

full_contacts_yeast()

contacts_yeast_eco1()

full_contacts_yeast()
}
\description{
Example datasets provided in \code{HiContacts} & \code{HiContactsData}
}
\examples{
contacts_yeast()
contacts_yeast_eco1()
full_contacts_yeast()
}
