% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmetics.R
\name{detrend}
\alias{detrend}
\title{detrend}
\usage{
detrend(x, use.scores = "balanced")
}
\arguments{
\item{x}{a \code{contacts} object}

\item{use.scores}{use.scores}
}
\value{
a \code{contacts} object with two additional scoress: \code{expected} and
\code{detrended}
}
\description{
detrend
}
\examples{
library(HiContacts)
data(contacts_yeast)
contacts_yeast <- detrend(contacts_yeast)
scores(contacts_yeast)
}
