% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diamond.R
\name{diamond}
\alias{diamond}
\alias{getDiamondInsulation}
\title{Contact map insulation}
\usage{
getDiamondInsulation(
  x,
  resolution = NULL,
  window_size = NULL,
  BPPARAM = BiocParallel::bpparam()
)
}
\arguments{
\item{x}{A \code{HiCExperiment} object over a full genome}

\item{resolution}{Which resolution to use to compute diamond insulation score}

\item{window_size}{Which window size to use to compute diamond insulation
score (default: 10 * resolution)}

\item{BPPARAM}{BiocParallel parallelization settings}

\item{weak_threshold}{Less stringent cutoff to call borders in the
diamond insulation score}

\item{strong_threshold}{More stringent cutoff to call borders in the
diamond insulation score}
}
\value{
a \code{HiCExperiment} object with additional \code{insulation} metadata,
containing the diamond insulation score computed
}
\description{
Computes diamond insulation score along the entire genome
}
\examples{
library(HiContacts)
full_contacts_yeast <- full_contacts_yeast()
diams <- getDiamondInsulation(full_contacts_yeast, resolution = 1000)
getDiamondInsulation(diams)
}
