% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contacts.R
\docType{methods}
\name{dim-contacts}
\alias{dim-contacts}
\alias{dim,contacts-method}
\title{dim method for objects of class \code{contacts}.}
\usage{
\S4method{dim}{contacts}(x)
}
\arguments{
\item{x}{A \code{contacts} object.}
}
\description{
dim method for objects of class \code{contacts}.
}
