% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{full_contacts_yeast}
\alias{full_contacts_yeast}
\title{Example of a \code{contacts} object}
\format{
An object of class \code{contacts} with 757 rows and 757 columns.
}
\source{
HiContactsData

fpath <- HiContactsData::HiContactsData('yeast_wt', 'mcool')
full_contacts_yeast <- contacts(fpath, resolution = 16000)
}
\usage{
full_contacts_yeast
}
\description{
Example of a \code{contacts} object
}
\keyword{datasets}
