% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{getAnchors}
\alias{getAnchors}
\title{This was adapted from \code{dovetail-genomics/coolR}}
\usage{
getAnchors(file, res = NULL, balanced = "cooler")
}
\description{
This was adapted from \code{dovetail-genomics/coolR}
}
