% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{getCounts}
\alias{getCounts}
\title{getCounts}
\usage{
getCounts(file, coords, anchors, coords2 = NULL, resolution = NULL)
}
\arguments{
\item{file}{file}

\item{coords}{coords}

\item{anchors}{anchors}

\item{coords2}{coords2}

\item{resolution}{resolution}
}
\description{
This was adapted from \code{dovetail-genomics/coolR}
}
