% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loops.R
\name{getLoops}
\alias{getLoops}
\title{Finding loops in contact map}
\usage{
getLoops(...)
}
\arguments{
\item{...}{Parameters passed to \code{HiCool::getLoops()}.}
}
\description{
Find loops using chromosight.

This function is actually provided by the \code{HiCool} package rather than
the \code{HiContacts} package. \code{HiCool} provides a self-managed
\code{conda} environment, and this limits
}
