% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ps.R
\name{getPs}
\alias{getPs}
\title{getPs}
\usage{
getPs(
  x,
  by_chr = FALSE,
  filtered_chr = c("XII", "chrXII", "chr12", "12", "Mito", "MT", "chrM")
)
}
\arguments{
\item{x}{x}

\item{by_chr}{by_chr}

\item{filtered_chr}{filtered_chr}
}
\description{
getPs
}
