% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg.R
\name{ggcorrmatrix}
\alias{ggcorrmatrix}
\title{ggcorrmatrix}
\usage{
ggcorrmatrix(mat, ticks = TRUE, cols, limits)
}
\arguments{
\item{mat}{mat}

\item{ticks}{ticks}

\item{cols}{cols}

\item{limits}{limits}
}
\description{
ggcorrmatrix
}
