% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ggMatrix}
\alias{ggMatrix}
\title{ggmatrix}
\usage{
ggMatrix(mat, ticks = TRUE, cols = afmhotr_colors, limits)
}
\arguments{
\item{mat}{mat}

\item{ticks}{ticks}

\item{cols}{cols}

\item{limits}{limits}
}
\description{
ggmatrix
}
