% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ggthemeHiContacts}
\alias{ggthemeHiContacts}
\title{ggplot2-related functions}
\usage{
ggthemeHiContacts(ticks = TRUE, grid = FALSE)
}
\arguments{
\item{ticks}{ticks}

\item{grid}{grid}
}
\value{
a custom ggplot2 theme
}
\description{
ggplot2-related functions
}
