% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{gi2mat}
\alias{gi2mat}
\title{gi2mat}
\usage{
gi2mat(gis, limits = NULL, truncate_tip = 0.5)
}
\arguments{
\item{gis}{gis}

\item{limits}{limits}

\item{truncate_tip}{truncate_tip}
}
\description{
gi2mat
}
