% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiContacts.R
\name{HiContacts package}
\alias{HiContacts package}
\title{HiContacts package}
\description{
HiContacts: R interface to (m)cool files and other Hi-C
processed file formats. HiContacts provides a collection of tools to
analyse and visualize Hi-C datasets. It can import data from pairs
or (m)cool files.
}
