% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{HiContacts-plots}
\alias{HiContacts-plots}
\title{HiContacts plotting functionalities}
\description{
Several plots can be generated in HiContacts:
\itemize{
\item Hi-C contact matrices
\item Distance-dependant interaction frequency decay (a.k.a. "Distance law" or "P(s)")
\item Virtual 4C profiles
\item Scalograms
\item Saddle plots
}
}
