% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contacts.R
\docType{methods}
\name{length-contacts}
\alias{length-contacts}
\alias{length,contacts-method}
\title{length method for objects of class \code{contacts}.}
\usage{
\S4method{length}{contacts}(x)
}
\arguments{
\item{x}{A \code{contacts} object.}
}
\description{
length method for objects of class \code{contacts}.
}
