% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{lsCoolResolutions}
\alias{lsCoolResolutions}
\title{lsCoolResolutions}
\usage{
lsCoolResolutions(file, silent = FALSE)
}
\arguments{
\item{file}{file}

\item{silent}{silent}
}
\description{
lsCoolResolutions
}
