% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmetics.R
\name{merge}
\alias{merge}
\title{merge}
\usage{
merge(..., use.scores = "balanced")
}
\arguments{
\item{...}{\code{contacts} objects}

\item{use.scores}{use.scores}
}
\value{
a \code{contacts} object. Each returned scores is the sum of the
corresponding scores from input \code{contacts}.
}
\description{
merge
}
\examples{
library(HiContacts)
data(contacts_yeast)
data(contacts_yeast_eco1)
merged_contacts <- merge(contacts_yeast_eco1, contacts_yeast)
merged_contacts
}
