% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{pairs2gi}
\alias{pairs2gi}
\title{pairs2gi}
\usage{
pairs2gi(pairs_file, columns = seq(2, 7), threads = 16)
}
\arguments{
\item{pairs_file}{pairs_file}

\item{columns}{columns}

\item{threads}{threads}
}
\description{
pairs2gi
}
