% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{getAnchors}
\alias{getAnchors}
\alias{getCountsFromPair}
\alias{getCounts}
\alias{fetchCool}
\alias{lsCoolFiles}
\alias{lsCoolResolutions}
\alias{peekCool}
\alias{cool2seqinfo}
\alias{cool2gi}
\alias{gi2cm}
\alias{cm2matrix}
\alias{pairs2gi}
\title{Parsing (m)cool files}
\usage{
getAnchors(file, resolution = NULL, balanced = "cooler")

getCountsFromPair(file, pair, anchors, resolution = NULL)

getCounts(file, coords, anchors, resolution = NULL)

fetchCool(file, path, resolution = NULL, idx = NULL, ...)

lsCoolFiles(file, verbose = FALSE)

lsCoolResolutions(file, verbose = FALSE)

peekCool(file, path, resolution = NULL, n = 10)

cool2seqinfo(file, resolution = NULL)

cool2gi(file, coords = NULL, resolution = NULL)

gi2cm(gi)

cm2matrix(cm, replace_NA = NA)

pairs2gi(
  file,
  chr1.field = 2,
  start1.field = 3,
  chr2.field = 4,
  start2.field = 5,
  nThread = 16,
  nrows = Inf
)
}
\arguments{
\item{file}{pairs file: \verb{<readname>\\t<chr1>\\t<start1>\\t<chr2>\\t<start2>}}

\item{resolution}{resolution}

\item{balanced}{import balancing scores}

\item{pair}{pair
(e.g. S4Vectors::Pairs(GRanges("II:200000-300000"), GRanges("II:70000-100000"))).}

\item{anchors}{anchors}

\item{coords}{NULL, character, or GRanges.
Can also be a Pairs object of paired GRanges (length of 1).}

\item{path}{path}

\item{idx}{idx to extract in cool file}

\item{...}{...}

\item{verbose}{Print resolutions in the console}

\item{n}{n}

\item{gi}{A \code{GInteractions} object}

\item{cm}{A \code{ContactMatrix} object}

\item{replace_NA}{Replace NA values}

\item{chr1.field}{chr1.field}

\item{start1.field}{start1.field}

\item{chr2.field}{chr2.field}

\item{start2.field}{start2.field}

\item{nThread}{Number of CPUs to use to import the \code{pairs} file in R}

\item{nrows}{Number of pairs to import}
}
\value{
anchors from (m)cool, stored as a GRanges

counts from (m)cool, stored as a tibble

counts from (m)cool, stored as a tibble

vector

vector

vector

vector

a Seqinfo object

a GenomicInteractions object

a ContactMatrix object

a dense matrix

a GenomicInteractions object
}
\description{
These functions are the workhorse internal functions used to import
a \verb{.(m)cool} file as GenomicInteractions (wrapped into a \code{contacts} object
by \code{contacts()} function).
}
