% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{peekCool}
\alias{peekCool}
\title{peekCool}
\usage{
peekCool(file, path, resolution = NULL, n = 10)
}
\arguments{
\item{file}{file}

\item{path}{path}

\item{resolution}{resolution}

\item{n}{n}
}
\description{
peekCool
}
