% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4C.R
\name{plot4C}
\alias{plot4C}
\title{plot4C}
\usage{
plot4C(x, mapping)
}
\arguments{
\item{x}{Output of virtual4C}

\item{mapping}{aes to pass on to ggplot2}
}
\value{
ggplot
}
\description{
plot4C
}
\examples{
library(HiContacts)
data(contacts_yeast)
v4C <- virtual4C(contacts_yeast, GenomicRanges::GRanges('II:490000-510000'))
plot4C(v4C, ggplot2::aes(x = center, y = score))
}
