% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotAggregatedMatrix}
\alias{plotAggregatedMatrix}
\title{plotAggregatedMatrix}
\usage{
plotAggregatedMatrix(
  file,
  coords,
  res = NULL,
  limits = NULL,
  dpi = 500,
  rasterize = TRUE,
  symmetrical = TRUE,
  BPPARAM = BiocParallel::bpparam()
)
}
\arguments{
\item{file}{file}

\item{coords}{coords}

\item{res}{res}

\item{limits}{limits}

\item{dpi}{dpi}

\item{rasterize}{rasterize}

\item{symmetrical}{symmetrical}

\item{BPPARAM}{BPPARAM}
}
\description{
plotAggregatedMatrix
}
