% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotCorrelatedMatrix}
\alias{plotCorrelatedMatrix}
\title{plotCorrelatedMatrix}
\usage{
plotCorrelatedMatrix(gis, limits = c(-1, 1), dpi = 500, rasterize = TRUE)
}
\arguments{
\item{gis}{gis}

\item{limits}{limits}

\item{dpi}{dpi}

\item{rasterize}{rasterize}
}
\description{
plotCorrelatedMatrix
}
