% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotMatrix}
\alias{plotMatrix}
\title{plotMatrix}
\usage{
plotMatrix(gis, limits = NULL, dpi = 500, rasterize = TRUE, symmetrical = TRUE)
}
\arguments{
\item{gis}{gis}

\item{limits}{limits}

\item{dpi}{dpi}

\item{rasterize}{rasterize}

\item{symmetrical}{symmetrical}
}
\description{
plotMatrix
}
