% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotMatrixList}
\alias{plotMatrixList}
\title{plotMatrixList}
\usage{
plotMatrixList(ls, limits = NULL, dist_max = 0.2)
}
\arguments{
\item{ls}{ls}

\item{limits}{limits}

\item{dist_max}{dist_max}
}
\description{
plotMatrixList
}
