% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotOverExpected}
\alias{plotOverExpected}
\title{plotOverExpected}
\usage{
plotOverExpected(
  gis,
  limits = c(-1, 1),
  dpi = 500,
  rasterize = TRUE,
  return_expected = FALSE,
  return_all = FALSE
)
}
\arguments{
\item{gis}{gis}

\item{limits}{limits}

\item{dpi}{dpi}

\item{rasterize}{rasterize}

\item{return_expected}{return_expected}

\item{return_all}{return_all}
}
\description{
plotOverExpected
}
