% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{contacts_yeast}
\alias{contacts_yeast_eco1}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{HiCExperiment}{\code{\link[HiCExperiment:data]{contacts_yeast}}, \code{\link[HiCExperiment:data]{contacts_yeast_eco1}}}
}}

