% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotScalogram}
\alias{plotScalogram}
\title{Plotting scalograms}
\usage{
plotScalogram(x, ylim = c(500, 1e+05))
}
\arguments{
\item{x}{GRanges, the output of \code{scalogram()}}

\item{ylim}{Range of distances to use for y-axis in scalograms}
}
\value{
ggplot
}
\description{
Plotting scalograms
}
\examples{
contacts_yeast <- HiCExperiment::contacts_yeast()
pairsFile(contacts_yeast) <- HiContactsData::HiContactsData(
  'yeast_wt', format = 'pairs.gz'
)
scalo <- scalogram(contacts_yeast['II'])
plotScalogram(scalo)
}
