% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contacts.R
\docType{methods}
\name{show-contacts}
\alias{show-contacts}
\alias{show,contacts-method}
\title{Show method for objects of class \code{contacts}.}
\usage{
\S4method{show}{contacts}(object)
}
\arguments{
\item{x}{A \code{contacts} object.}
}
\description{
Show method for objects of class \code{contacts}.
}
