% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShREC.R
\name{ShREC}
\alias{ShREC}
\title{ShREC}
\usage{
ShREC(x, use.scores = "count", alpha = 1, min.nnz = 50)
}
\arguments{
\item{x}{A \code{HiCExperiment} object}

\item{use.scores}{Which scores to use to perform operations}

\item{alpha}{Power law scaling factor. As indicated in Boost-HiC documentation,
the alpha parameter influences the weighting of contacts: if alpha < 1
long-range interactions are prioritized; if alpha >> 1 short-range
interactions have more weight when computing the distance matrix}

\item{min.nnz}{Minimal number of non-zero elements per row used to mask
lowly-covered bins}
}
\value{
A \code{HiCExperiment} object with additional scores
}
\description{
ShREC algorithm allows one to model 3D structure of individual chromosomes
very fast. This algorithm has been proposed in
\href{https://doi.org/10.1038/nmeth.3104}{Lesne et al., 2014} and is implemented
in R here.
}
\examples{
library(HiContacts)
full_contacts_yeast <- full_contacts_yeast()
boost(full_contacts_yeast)
}
