% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmetics.R
\name{smoothen}
\alias{smoothen}
\title{smoothen}
\usage{
smoothen(
  x,
  use.assay = "balanced",
  use_serpentine_trend = TRUE,
  serpentine_niter = 10L,
  serpentine_ncores = 16L
)
}
\description{
smoothen
}
