% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracks.R
\name{tracks}
\alias{tracks}
\alias{alignTrack}
\title{Aligning tracks with contact maps}
\usage{
alignTrack(x, bw, track.name = "track")
}
\arguments{
\item{x}{A \code{HiCExperiment} object over a full genome}

\item{bw}{a bigwig track imported as 'Rle'}

\item{track.name}{Name to give to the track in regions metadata}
}
\value{
A \code{HiCExperiment} object with 2 added columns in \code{regions(x)}
}
\description{
Aligning tracks with contact maps
}
