% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{splitCoords}
\alias{splitCoords}
\alias{formatCoords}
\alias{char2pair}
\alias{getHicStats}
\alias{fullContactInteractions}
\alias{sdiag<-}
\alias{sortPairs}
\alias{asGInteractions}
\title{Utils functions}
\usage{
splitCoords(coords)

formatCoords(coords)

char2pair(char)

getHicStats(hicstuff_log)

fullContactInteractions(chr, start, end, binning)

sdiag(A, k = 0) <- value

sortPairs(pairs)

asGInteractions(df)
}
\arguments{
\item{coords}{coords}

\item{char}{char (e.g. "II:30000-50000" or "II:30000-50000 x II:60000-80000")}

\item{hicstuff_log}{log file generated by hicstuff}

\item{chr}{chr}

\item{start}{start}

\item{end}{end}

\item{binning}{binning}

\item{A}{A}

\item{k}{k}

\item{value}{value}

\item{pairs}{pairs}

\item{df}{df}
}
\value{
a list containing \code{chr}, \code{start} and \code{end}

a character string

a S4Vectors::Pairs object

a list

a GenomicInteractions object

a matrix

a Pairs object

a GenomicInteractions object
}
\description{
Utilities to facilitate parsing/handling of coordinates, GInteractions,
Pairs, ...
}
