% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APA.R
\name{APA}
\alias{APA}
\title{APA}
\usage{
APA(
  x,
  coords,
  bins = 50,
  use.assay = "balanced",
  BPPARAM = BiocParallel::bpparam()
)
}
\arguments{
\item{x}{x}

\item{coords}{coords}

\item{BPPARAM}{BPPARAM}

\item{res}{res}

\item{limits}{limits}

\item{dpi}{dpi}

\item{rasterize}{rasterize}

\item{symmetrical}{symmetrical}

\item{scale}{scale}

\item{cmap}{cmap}
}
\description{
APA
}
