% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APA.R
\name{APA}
\alias{APA}
\title{APA}
\usage{
APA(
  x,
  coords,
  bins = 50,
  use.scores = "balanced",
  BPPARAM = BiocParallel::bpparam()
)
}
\arguments{
\item{x}{x}

\item{coords}{coords}

\item{bins}{bins}

\item{use.scores}{use.scores}

\item{BPPARAM}{BPPARAM}
}
\value{
an aggregated contact matrix
}
\description{
APA
}
