% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APA.R
\name{APA}
\alias{APA}
\title{Aggregated Plot Analysis}
\usage{
APA(x, coords, nbins = 50, use.scores = "balanced")
}
\arguments{
\item{x}{x}

\item{coords}{coords}

\item{use.scores}{use.scores}

\item{bins}{bins}
}
\value{
an aggregated contact matrix
}
\description{
Aggregated Plot Analysis
}
