% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R, R/arithmetics.R
\name{aggregate,HiCExperiment-method}
\alias{aggregate,HiCExperiment-method}
\alias{arithmetics}
\alias{detrend}
\alias{autocorrelate}
\alias{divide}
\alias{merge}
\alias{despeckle}
\title{HiContacts arithmetics functionalities}
\usage{
\S4method{aggregate}{HiCExperiment}(x, ...)

detrend(x, use.scores = "balanced")

autocorrelate(x, use.scores = "balanced", detrend = TRUE, ignore_ndiags = 3)

divide(x, by, use.scores = "balanced")

merge(..., use.scores = "balanced")

despeckle(x, use.scores = "balanced", focal.size = 1)
}
\arguments{
\item{x}{a \code{HiCExperiment} object}

\item{...}{\code{HiCExperiment} objects. For \code{aggregate}, \code{targets} (a set of
GRanges or GInteractions).}

\item{use.scores}{Which scores to use to perform operations}

\item{detrend}{Detrend matrix before performing autocorrelation}

\item{ignore_ndiags}{ignore N diagonals when calculating correlations}

\item{by}{a \code{HiCExperiment} object}

\item{focal.size}{Size of the smoothing rectangle}
}
\value{
a \code{HiCExperiment} object with extra scores
}
\description{
Different arithmetic operations can be performed:
\itemize{
\item Detrending a contact matrix, i.e. removing the distance-dependent
contact trend;
\item Autocorrelate a contact matrix: this is is typically done to highlight
large-scale compartments;
\item Divide one contact matrix by another;
\item Merge multiple contact matrices;
\item Despeckle (i.e. smooth out) a contact matrix out;
\item Aggregate (average) a contact matrices over a set of genomic loci of
interest.
}
}
\examples{
#### -----
#### Detrending a contact matrix
#### -----

library(HiContacts)
contacts_yeast <- contacts_yeast()
contacts_yeast <- detrend(contacts_yeast)
contacts_yeast

#### -----
#### Auto-correlate a contact matrix
#### -----

autocorrelate(contacts_yeast)

#### -----
#### Divide 2 contact matrices
#### -----

contacts_yeast <- refocus(contacts_yeast, 'II')
contacts_yeast_eco1 <- contacts_yeast_eco1() |> refocus('II')
divide(contacts_yeast_eco1, by = contacts_yeast)

#### -----
#### Merge 2 contact matrices
#### -----

merge(contacts_yeast_eco1, contacts_yeast)

#### -----
#### Despeckle (smoothen) a contact map
#### -----

despeckle(contacts_yeast)

#### -----
#### Aggregate a contact matrix over centromeres, at different scales
#### -----

contacts <- contacts_yeast() |> zoom(resolution = 1000)
centros <- topologicalFeatures(contacts, 'centromeres')
aggregate(contacts, targets = centros, flanking_bins = 50)

}
