% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmetics.R
\name{detrend}
\alias{detrend}
\alias{autocorrelate}
\alias{divide}
\alias{merge}
\alias{serpentinify}
\title{Arithmetics with (m)cool file(s)}
\usage{
detrend(x, use.scores = "balanced")

autocorrelate(x, use.scores = "balanced", detrend = TRUE, ignore_ndiags = 3)

divide(x, by, use.scores = "balanced")

merge(..., use.scores = "balanced")

serpentinify(
  x,
  use.scores = "balanced",
  use_serpentine_trend = TRUE,
  serpentine_niter = 10L,
  serpentine_ncores = 16L
)
}
\arguments{
\item{x}{a \code{HiCExperiment} object}

\item{use.scores}{use.scores}

\item{ignore_ndiags}{ignore N diagonals when calculating correlations}

\item{by}{a \code{HiCExperiment} object}

\item{...}{\code{HiCExperiment} objects}

\item{use_serpentine_trend}{whether to use the trend estimated with
serpentine (this requires \code{reticulate} and the python package \code{serpentine})}

\item{serpentine_niter}{number of iterations to use for serpentine}

\item{serpentine_ncores}{number of CPUs to use for serpentine}
}
\value{
a \code{HiCExperiment} object with two additional scoress: \code{expected} and
\code{detrended}

a \code{HiCExperiment} object with a single \code{autocorrelation} scores

a \code{HiCExperiment} object with a single \code{ratio} scores

a \code{HiCExperiment} object. Each returned scores is the sum of the
corresponding scores from input \code{HiCExperiment}.

a \code{HiCExperiment} object with a single \code{smoothen} scores
}
\description{
Different operations can be performed:
\itemize{
\item Detrending a contact matrix, i.e. removing the distance-dependent
contact trend;
\item Autocorrelate a contact matrix: this is is typically done to highlight
large-scale compartments;
\item Divide one contact matrix by another;
\item Merge multiple contact matrices;
\item Serpentinify, or smooth a contact matrix out. This requires \code{serpentine}
python package to be installed.
}
}
\examples{
#### -----
#### Detrending a contact matrix
#### -----

library(HiContacts)
contacts_yeast <- contacts_yeast()
contacts_yeast <- detrend(contacts_yeast)
scores(contacts_yeast)
#### -----
#### Auto-correlate a contact matrix
#### -----

contacts_yeast <- autocorrelate(contacts_yeast)
scores(contacts_yeast)
plotMatrix(contacts_yeast, scale = 'linear', limits = c(-1, 1), cmap = bwrColors())
#### -----
#### Divide 2 contact matrices
#### -----

contacts_yeast <- contacts_yeast()
contacts_yeast_eco1 <- contacts_yeast_eco1()
div_contacts <- divide(contacts_yeast_eco1, by = contacts_yeast)
div_contacts
plotMatrix(div_contacts, scale = 'log2', limits = c(-2, 2), cmap = bwrColors())
#### -----
#### Merge 2 contact matrices
#### -----

merged_contacts <- merge(contacts_yeast_eco1, contacts_yeast)
merged_contacts
}
