% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R, R/arithmetics.R
\name{arithmetics}
\alias{arithmetics}
\alias{aggregate,HiCExperiment-method}
\alias{detrend}
\alias{autocorrelate}
\alias{divide}
\alias{merge}
\alias{despeckle}
\title{Arithmetics with (m)cool file(s)}
\usage{
\S4method{aggregate}{HiCExperiment}(x, ...)

detrend(x, use.scores = "balanced")

autocorrelate(x, use.scores = "balanced", detrend = TRUE, ignore_ndiags = 3)

divide(x, by, use.scores = "balanced")

merge(..., use.scores = "balanced")

despeckle(x, use.scores = "balanced", focal.size = 1)
}
\arguments{
\item{x}{a \code{HiCExperiment} object}

\item{...}{\code{HiCExperiment} objects. For \code{aggregate}, \code{targets} (a set of
GRanges or GInteractions).}

\item{use.scores}{use.scores}

\item{detrend}{Detrend matrix before performing autocorrelation}

\item{ignore_ndiags}{ignore N diagonals when calculating correlations}

\item{by}{a \code{HiCExperiment} object}

\item{focal.size}{Size of the smoothing rectangle}
}
\value{
a \code{AggrHiCExperiment} object

a \code{HiCExperiment} object with two additional scoress: \code{expected} and
\code{detrended}

a \code{HiCExperiment} object with a single \code{autocorrelation} scores

a \code{HiCExperiment} object with a single \code{ratio} scores

a \code{HiCExperiment} object. Each returned scores is the sum of the
corresponding scores from input \code{HiCExperiment}.

a \code{HiCExperiment} object with a single \code{smoothen} scores
}
\description{
Different operations can be performed:
\itemize{
\item Detrending a contact matrix, i.e. removing the distance-dependent
contact trend;
\item Autocorrelate a contact matrix: this is is typically done to highlight
large-scale compartments;
\item Divide one contact matrix by another;
\item Merge multiple contact matrices;
\item Despeckle (i.e. smooth out) a contact matrix out.
\item Aggregate (average) a contact matrices over a set of genomic loci of
interest;
}
}
\examples{
#### -----
#### Aggregate a contact matrix over centromeres, at different scales
#### -----

contacts <- full_contacts_yeast() |> zoom(resolution = 1000)
centros <- topologicalFeatures(contacts, 'centromeres')
aggr <- aggregate(contacts, targets = centros, flanking_bins = 50)
plotMatrix(aggr, 'detrended', scale = 'linear', limits = c(-1, 1))

contacts <- full_contacts_yeast() |> zoom(resolution = 8000)
centros <- topologicalFeatures(contacts, 'centromeres')
aggr <- aggregate(contacts, targets = centros, flanking_bins = 20)
plotMatrix(aggr, 'detrended', scale = 'linear', limits = c(-1, 1))

#### -----
#### Detrending a contact matrix
#### -----

library(HiContacts)
contacts_yeast <- contacts_yeast()
contacts_yeast <- detrend(contacts_yeast)
scores(contacts_yeast)

#### -----
#### Auto-correlate a contact matrix
#### -----

contacts_yeast <- autocorrelate(contacts_yeast)
scores(contacts_yeast)
plotMatrix(contacts_yeast, scale = 'linear', limits = c(-1, 1), cmap = bwrColors())

#### -----
#### Divide 2 contact matrices
#### -----

contacts_yeast <- contacts_yeast() |> refocus('II')
contacts_yeast_eco1 <- contacts_yeast_eco1() |> refocus('II')
div_contacts <- divide(contacts_yeast_eco1, by = contacts_yeast)
div_contacts
plotMatrix(div_contacts, scale = 'log2', limits = c(-2, 2), cmap = bwrColors())

#### -----
#### Merge 2 contact matrices
#### -----

merged_contacts <- merge(contacts_yeast_eco1, contacts_yeast)
merged_contacts

#### -----
#### Despeckle (smoothen) a contact map
#### -----

smoothed_contacts <- despeckle(contacts_yeast)
smoothed_contacts

}
