% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmetics.R
\name{autocorrelate}
\alias{autocorrelate}
\title{autocorrelate}
\usage{
autocorrelate(x, use.assay = "balanced", ignore_ndiags = 3)
}
\arguments{
\item{x}{a \code{contacts} object}

\item{use.assay}{use.assay}

\item{ignore_ndiags}{ignore N diagonals when calculating correlations}
}
\value{
a \code{contacts} object with a single \code{autocorrelation} assay
}
\description{
autocorrelate
}
\examples{
library(HiContacts)
data(contacts_yeast)
contacts_yeast <- autocorrelate(contacts_yeast)
assays(contacts_yeast)
plotMatrix(contacts_yeast, scale = 'linear', limits = c(-1, 1), cmap = bwr_colors)
}
