% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_cool_file}
\alias{check_cool_file}
\alias{check_resolution}
\alias{check_cool_format}
\alias{is_mcool}
\alias{is_cool}
\alias{is_same_seqinfo}
\alias{is_same_resolution}
\alias{is_same_bins}
\alias{is_same_regions}
\alias{is_comparable}
\alias{is_square}
\alias{is_symmetrical}
\title{Checks functions}
\usage{
check_cool_file(path)

check_resolution(contacts, resolution)

check_cool_format(path, resolution)

is_mcool(path)

is_cool(path)

is_same_seqinfo(...)

is_same_resolution(...)

is_same_bins(...)

is_same_regions(...)

is_comparable(...)

is_square(pair)

is_symmetrical(contacts)
}
\arguments{
\item{path}{Path of a (m)cool file}

\item{contacts}{A \code{contacts} object}

\item{resolution}{Resolution}

\item{...}{\code{contacts} object}

\item{pair}{Pairs object with length of 1}
}
\value{
Logical
}
\description{
Useful functions to validate the nature/structure of (m)cool files or
\code{contacts} objects.
}
\examples{
library(HiContacts)
contacts_yeast <- contacts_yeast()
is_symmetrical(contacts_yeast)
}
