% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotMatrix}
\alias{plotMatrix}
\alias{ggMatrix}
\alias{ggHorizontalMatrix}
\title{Plotting a contact matrix}
\usage{
plotMatrix(
  x,
  use.scores = NULL,
  scale = "log10",
  max.distance = NULL,
  loops = NULL,
  borders = NULL,
  limits = NULL,
  dpi = 500,
  rasterize = TRUE,
  symmetrical = TRUE,
  chrom_lines = TRUE,
  show_grid = FALSE,
  cmap = NULL
)

ggMatrix(mat, ticks = TRUE, grid = FALSE, cols = coolerColors(), limits)

ggHorizontalMatrix(df, cols = coolerColors(), limits)
}
\arguments{
\item{x}{x}

\item{use.scores}{use.scores}

\item{scale}{scale}

\item{max.distance}{maximum distance. If provided, the heatmap is plotted
horizontally.}

\item{loops}{loops}

\item{borders}{borders}

\item{limits}{limits}

\item{dpi}{dpi}

\item{rasterize}{rasterize}

\item{symmetrical}{symmetrical}

\item{chrom_lines}{chrom_lines}

\item{show_grid}{show_grid}

\item{cmap}{color map}

\item{mat, df}{mat}

\item{ticks}{ticks}

\item{grid}{grid}

\item{cols}{cols}
}
\value{
ggplot object

ggplot
}
\description{
Plotting a contact matrix
}
\examples{
library(HiContacts)
contacts_yeast <- contacts_yeast()
plotMatrix(
    contacts_yeast, 
    use.scores = 'balanced', 
    scale = 'log10', 
    limits = c(-4, -1)
)
}
