% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contacts.R
\docType{class}
\name{Contacts-class}
\alias{Contacts-class}
\alias{Contacts}
\alias{fileName}
\alias{fileName,Contacts-method}
\alias{resolutions}
\alias{resolutions,Contacts-method}
\alias{resolution}
\alias{resolution,Contacts-method}
\alias{focus}
\alias{focus,Contacts-method}
\alias{focus<-}
\alias{focus<-,Contacts,character-method}
\alias{interactions}
\alias{interactions,Contacts-method}
\alias{interactions<-}
\alias{interactions<-,Contacts,GInteractions-method}
\alias{scores}
\alias{scores,Contacts,missing-method}
\alias{scores,Contacts,character-method}
\alias{scores,Contacts,numeric-method}
\alias{scores<-}
\alias{scores<-,Contacts,character,numeric-method}
\alias{topologicalFeatures}
\alias{topologicalFeatures,Contacts,missing-method}
\alias{topologicalFeatures,Contacts,character-method}
\alias{topologicalFeatures,Contacts,numeric-method}
\alias{topologicalFeatures<-}
\alias{topologicalFeatures<-,Contacts,character,GRangesOrGInteractions-method}
\alias{pairsFile}
\alias{pairsFile,Contacts-method}
\alias{pairsFile<-}
\alias{pairsFile<-,Contacts,character-method}
\alias{metadata<-}
\alias{metadata<-,Contacts,list-method}
\alias{length}
\alias{length,Contacts-method}
\alias{[}
\alias{[,Contacts,numeric,ANY,ANY-method}
\alias{[,Contacts,logical,ANY,ANY-method}
\alias{[,Contacts,character,ANY,ANY-method}
\alias{seqinfo,Contacts-method}
\alias{bins}
\alias{bins,Contacts-method}
\alias{anchors}
\alias{anchors,Contacts-method}
\alias{regions}
\alias{regions,Contacts-method}
\alias{summary}
\alias{summary,Contacts-method}
\alias{show}
\alias{show,Contacts-method}
\alias{setAs}
\alias{setAs,Contacts-method}
\title{\code{Contacts} S4 class and methods}
\usage{
Contacts(
  file,
  resolution = NULL,
  focus = NULL,
  metadata = list(),
  topologicalFeatures = S4Vectors::SimpleList(loops =
    S4Vectors::Pairs(GenomicRanges::GRanges(), GenomicRanges::GRanges()), borders =
    GenomicRanges::GRanges(), compartments = GenomicRanges::GRanges(), viewpoints =
    GenomicRanges::GRanges()),
  pairsFile = NULL
)

\S4method{fileName}{Contacts}(object)

\S4method{resolutions}{Contacts}(x)

\S4method{resolution}{Contacts}(x)

\S4method{focus}{Contacts}(x)

\S4method{focus}{Contacts,character}(x) <- value

\S4method{interactions}{Contacts}(x)

\S4method{interactions}{Contacts,GInteractions}(x) <- value

\S4method{scores}{Contacts,missing}(x)

\S4method{scores}{Contacts,character,numeric}(x, name) <- value

\S4method{topologicalFeatures}{Contacts,missing}(x)

\S4method{topologicalFeatures}{Contacts,character,GRangesOrGInteractions}(x, name) <- value

\S4method{pairsFile}{Contacts}(x)

\S4method{pairsFile}{Contacts,character}(x) <- value

\S4method{metadata}{Contacts,list}(x) <- value

\S4method{length}{Contacts}(x)

\S4method{[}{Contacts,numeric,ANY,ANY}(x, i)

\S4method{seqinfo}{Contacts}(x)

\S4method{bins}{Contacts}(x)

\S4method{anchors}{Contacts}(x)

\S4method{regions}{Contacts}(x)

\S4method{summary}{Contacts}(object)

\S4method{show}{Contacts}(object)
}
\arguments{
\item{file}{Path to a (m)cool file}

\item{resolution}{Resolution to use with mcool file}

\item{focus}{focus Chr. coordinates for which
interaction counts are extracted from the .(m)cool file, provided
as a character string (e.g. "II:4000-5000"). If not provided,
the entire (m)cool file will be imported.}

\item{metadata}{list of metadata}

\item{topologicalFeatures}{topologicalFeatures provided as a named SimpleList}

\item{pairsFile}{Path to an associated .pairs file}

\item{object}{A \code{Contacts} object.}

\item{x}{A \code{Contacts} object.}

\item{value}{value}

\item{name}{name}

\item{i}{a range or boolean vector.}
}
\value{
a new \code{Contacts} object.
}
\description{
\code{Contacts} S4 class and methods
}
\section{Slots}{

\describe{
\item{\code{fileName}}{Path of (m)cool file}

\item{\code{focus}}{Chr. coordinates for which interaction counts are extracted
from the .(m)cool file.}

\item{\code{resolutions}}{Resolutions available in the .(m)cool file.}

\item{\code{resolution}}{Current resolution}

\item{\code{interactions}}{Genomic Interactions extracted from the .(m)cool object}

\item{\code{scores}}{Available interaction scores.}

\item{\code{topologicalFeatures}}{Topological features associated with the dataset
(e.g. loops (\<Pairs\>), borders (\<GRanges\>),
viewpoints (\<GRanges\>), etc...)}

\item{\code{pairsFile}}{Path to the .pairs file associated with the .(m)cool file}

\item{\code{metadata}}{metadata associated with the .(m)cool file.}
}}

\examples{
library(HiContacts)
contacts_yeast <- contacts_yeast()
contacts_yeast
fileName(contacts_yeast)
resolutions(contacts_yeast)
resolution(contacts_yeast)
focus(contacts_yeast)
interactions(contacts_yeast)
scores(contacts_yeast)
tail(scores(contacts_yeast, 1))
tail(scores(contacts_yeast, 'balanced'))
scores(contacts_yeast, 'test') <- runif(length(contacts_yeast))
tail(scores(contacts_yeast, 'test'))
full_contacts_yeast <- full_contacts_yeast()
topologicalFeatures(full_contacts_yeast)
topologicalFeatures(full_contacts_yeast, 1)
topologicalFeatures(full_contacts_yeast, 'centromeres')
data(centros_yeast)
topologicalFeatures(contacts_yeast, 'centromeres') <- centros_yeast
topologicalFeatures(contacts_yeast, 'centromeres')
pairsFile(full_contacts_yeast)
length(contacts_yeast)
contacts_yeast[seq_len(10)]
seqinfo(contacts_yeast)
bins(contacts_yeast)
anchors(contacts_yeast)
regions(contacts_yeast)
summary(contacts_yeast)
show(contacts_yeast)
as(contacts_yeast, 'GInteractions')
as(contacts_yeast, 'ContactMatrix')
as(contacts_yeast, 'matrix')[seq_len(10), seq_len(10)]
as(contacts_yeast, 'data.frame')
}
