% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contacts.R
\name{Contacts}
\alias{Contacts}
\title{Contacts}
\usage{
Contacts(
  file,
  resolution = NULL,
  focus = NULL,
  metadata = list(),
  topologicalFeatures = S4Vectors::SimpleList(loops =
    S4Vectors::Pairs(GenomicRanges::GRanges(), GenomicRanges::GRanges()), borders =
    GenomicRanges::GRanges(), compartments = GenomicRanges::GRanges(), viewpoints =
    GenomicRanges::GRanges()),
  pairsFile = NULL
)
}
\arguments{
\item{file}{Path to a (m)cool file}

\item{resolution}{Resolution to use with mcool file}

\item{focus}{focus Chr. coordinates for which
interaction counts are extracted from the .(m)cool file, provided
as a character string (e.g. "II:4001-5000"). If not provided,
the entire (m)cool file will be imported.}

\item{metadata}{list of metadata}

\item{topologicalFeatures}{topologicalFeatures provided as a named SimpleList}

\item{pairsFile}{Path to an associated .pairs file}
}
\value{
a new \code{HiCExperiment} object.
}
\description{
This function has been deprecated in favor of the generic \code{HiCExperiment()}
constructor (from HiCExperiment package).
}
\examples{
library(HiContacts)
library(HiContactsData)
mcool_path <- HiContactsData::HiContactsData('yeast_wt', 'mcool')
Contacts(mcool_path, resolution = 1000)
}
