% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contacts.R
\docType{methods}
\name{contacts}
\alias{contacts}
\alias{length}
\alias{length,contacts-method}
\alias{dim}
\alias{dim,contacts-method}
\alias{[}
\alias{[,contacts-method}
\alias{show}
\alias{show,contacts-method}
\title{contacts}
\usage{
contacts(
  cool_path,
  resolution = NULL,
  focus = NULL,
  metadata = NULL,
  features = NULL,
  pairs = NULL
)

\S4method{length}{contacts}(x)

\S4method{dim}{contacts}(x)

\S4method{[}{contacts,ANY,ANY,ANY}(x, i)

\S4method{show}{contacts}(object)
}
\arguments{
\item{cool_path}{Path of a (m)cool file}

\item{resolution}{Resolution to use with mcool file}

\item{focus}{focus Chr. coordinates for which
interaction counts are extracted from the .(m)cool file.
Can be provided as a character string or as a GRanges object}

\item{metadata}{list of metadata}

\item{features}{features provided as a named SimpleList}

\item{pairs}{Path to an associated .pairs file}

\item{x}{A \code{contacts} object.}

\item{i}{a range or boolean vector.}

\item{object}{A \code{contacts} object.}
}
\value{
a new \code{contacts} object.
}
\description{
contacts

length method for objects of class \code{contacts}.

dim method for objects of class \code{contacts}.

[ method for objects of class \code{contacts}.

Show method for objects of class \code{contacts}.
}
\examples{
library(HiContacts)
data(contacts_yeast)
contacts_yeast
library(HiContacts)
data(contacts_yeast)
length(contacts_yeast)
library(HiContacts)
data(contacts_yeast)
dim(contacts_yeast)
library(HiContacts)
data(contacts_yeast)
contacts_yeast[1:10]
library(HiContacts)
data(contacts_yeast)
contacts_yeast
}
