% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contacts.R
\docType{class}
\name{contacts-class}
\alias{contacts-class}
\alias{contacts}
\alias{length}
\alias{length,contacts-method}
\alias{[}
\alias{[,contacts,numeric,ANY,ANY-method}
\alias{[,contacts,logical,ANY,ANY-method}
\alias{[,contacts,character,ANY,ANY-method}
\alias{fileName}
\alias{fileName,contacts-method}
\alias{seqinfo,contacts-method}
\alias{resolutions}
\alias{resolutions,contacts-method}
\alias{resolution}
\alias{resolution,contacts-method}
\alias{bins}
\alias{bins,contacts-method}
\alias{focus}
\alias{focus,contacts-method}
\alias{interactions}
\alias{interactions,contacts-method}
\alias{interactions<-}
\alias{interactions<-,contacts,GInteractions-method}
\alias{scores}
\alias{scores,contacts,missing-method}
\alias{scores,contacts,character-method}
\alias{scores,contacts,numeric-method}
\alias{scores<-}
\alias{scores<-,contacts,character,numeric-method}
\alias{topologicalFeatures}
\alias{topologicalFeatures,contacts,missing-method}
\alias{topologicalFeatures,contacts,character-method}
\alias{topologicalFeatures,contacts,numeric-method}
\alias{topologicalFeatures<-}
\alias{topologicalFeatures<-,contacts,character,GRangesOrGInteractions-method}
\alias{pairsFile}
\alias{pairsFile,contacts-method}
\alias{pairsFile<-}
\alias{pairsFile<-,contacts,character-method}
\alias{anchors}
\alias{anchors,contacts-method}
\alias{regions}
\alias{regions,contacts-method}
\alias{summary}
\alias{summary,contacts-method}
\alias{show}
\alias{show,contacts-method}
\alias{setAs}
\alias{setAs,contacts-method}
\title{\code{contacts} S4 class and methods}
\usage{
contacts(
  file,
  resolution = NULL,
  focus = NULL,
  metadata = list(),
  topologicalFeatures = S4Vectors::SimpleList(loops =
    S4Vectors::Pairs(GenomicRanges::GRanges(), GenomicRanges::GRanges()), borders =
    GenomicRanges::GRanges(), compartments = GenomicRanges::GRanges(), viewpoints =
    GenomicRanges::GRanges()),
  pairsFile = NULL
)

\S4method{length}{contacts}(x)

\S4method{[}{contacts,numeric,ANY,ANY}(x, i)

\S4method{fileName}{contacts}(object)

\S4method{seqinfo}{contacts}(x)

resolutions(x)

\S4method{resolution}{contacts}(x)

bins(x)

focus(x)

\S4method{interactions}{contacts}(x)

\S4method{interactions}{contacts,GInteractions}(x) <- value

scores(x, name)

scores(x, name) <- value

topologicalFeatures(x, name)

topologicalFeatures(x, name) <- value

pairsFile(x, name)

pairsFile(x) <- value

\S4method{anchors}{contacts}(x)

\S4method{regions}{contacts}(x)

\S4method{summary}{contacts}(object)

\S4method{show}{contacts}(object)
}
\arguments{
\item{file}{Path to a (m)cool file}

\item{resolution}{Resolution to use with mcool file}

\item{focus}{focus Chr. coordinates for which
interaction counts are extracted from the .(m)cool file, provided
as a character string (e.g. "II:4000-5000"). If not provided,
the entire (m)cool file will be imported.}

\item{metadata}{list of metadata}

\item{topologicalFeatures}{topologicalFeatures provided as a named SimpleList}

\item{pairsFile}{Path to an associated .pairs file}

\item{x}{A \code{contacts} object.}

\item{i}{a range or boolean vector.}

\item{object}{A \code{contacts} object.}

\item{value}{value}

\item{name}{name}
}
\value{
a new \code{contacts} object.
}
\description{
\code{contacts} S4 class and methods
}
\section{Slots}{

\describe{
\item{\code{fileName}}{Path of (m)cool file}

\item{\code{focus}}{Chr. coordinates for which interaction counts are extracted
from the .(m)cool file.}

\item{\code{resolutions}}{Resolutions available in the .(m)cool file.}

\item{\code{resolution}}{Current resolution}

\item{\code{interactions}}{Genomic Interactions extracted from the .(m)cool object}

\item{\code{scores}}{Available interaction scores.}

\item{\code{topologicalFeatures}}{Topological features associated with the dataset
(e.g. loops (\<Pairs\>), borders (\<GRanges\>),
viewpoints (\<GRanges\>), etc...)}

\item{\code{pairsFile}}{Path to the .pairs file associated with the .(m)cool file}

\item{\code{metadata}}{metadata associated with the .(m)cool file.}
}}

\examples{
library(HiContacts)
contacts_yeast <- contacts_yeast()
contacts_yeast
length(contacts_yeast)
contacts_yeast[seq_len(10)]
fileName(contacts_yeast)
seqinfo(contacts_yeast)
resolutions(contacts_yeast)
resolution(contacts_yeast)
bins(contacts_yeast)
focus(contacts_yeast)
interactions(contacts_yeast)
scores(contacts_yeast)
scores(contacts_yeast, 1)
scores(contacts_yeast, 'balanced')
scores(contacts_yeast, 'test') <- runif(length(contacts_yeast))
scores(contacts_yeast, 'test')
full_contacts_yeast <- full_contacts_yeast()
topologicalFeatures(full_contacts_yeast)
topologicalFeatures(full_contacts_yeast, 1)
topologicalFeatures(full_contacts_yeast, 'centromeres')
data(centros_yeast)
topologicalFeatures(contacts_yeast, 'centromeres') <- centros_yeast
topologicalFeatures(contacts_yeast, 'centromeres')
pairsFile(full_contacts_yeast)
anchors(contacts_yeast)
regions(contacts_yeast)
summary(contacts_yeast)
show(contacts_yeast)
as(contacts_yeast, 'GInteractions')
as(contacts_yeast, 'ContactMatrix')
as(contacts_yeast, 'matrix')[seq_len(10), seq_len(10)]
}
