% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{cool2gi}
\alias{cool2gi}
\title{cool2gi}
\usage{
cool2gi(file, coords = NULL, coords2 = NULL, res = NULL)
}
\arguments{
\item{file}{file}

\item{coords}{coords}

\item{coords2}{coords2}

\item{res}{res}
}
\description{
cool2gi
}
