% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{centros_yeast}
\alias{centros_yeast}
\alias{contacts_yeast}
\alias{contacts_yeast_eco1}
\alias{full_contacts_yeast}
\title{Example datasets provided in \code{HiContacts} & \code{HiContactsData}}
\format{
An object of class \code{"GRanges"}.

An object of class \code{"Contacts"}.
}
\source{
HiContacts
}
\usage{
data(centros_yeast)

contacts_yeast()

contacts_yeast_eco1()

full_contacts_yeast()
}
\description{
Example datasets provided in \code{HiContacts} & \code{HiContactsData}
}
\examples{
data(centros_yeast)
centros_yeast
contacts_yeast()
contacts_yeast_eco1()
full_contacts_yeast()
}
\keyword{datasets}
