% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fetchCool}
\alias{fetchCool}
\title{fetchCool}
\usage{
fetchCool(file, path, res = NULL, idx = NULL, ...)
}
\arguments{
\item{file}{file}

\item{path}{path}

\item{res}{res}

\item{idx}{idx}

\item{...}{...}
}
\description{
fetchCool
}
