% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{getAnchors}
\alias{getAnchors}
\title{getAnchors}
\usage{
getAnchors(file, resolution = NULL, balanced = "cooler")
}
\arguments{
\item{file}{file}

\item{resolution}{resolution}

\item{balanced}{balanced}
}
\description{
getAnchors
}
