% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{getAnchors}
\alias{getAnchors}
\title{getAnchors}
\usage{
getAnchors(file, res = NULL, balanced = "cooler")
}
\arguments{
\item{file}{file}

\item{res}{res}

\item{balanced}{balanced}
}
\description{
This was adapted from \code{dovetail-genomics/coolR}
}
