% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg.R
\name{ggmatrix}
\alias{ggmatrix}
\title{ggmatrix}
\usage{
ggmatrix(mat, ticks = TRUE, cols = afmhotr_colors, limits)
}
\arguments{
\item{mat}{mat}

\item{ticks}{ticks}

\item{cols}{cols}

\item{limits}{limits}
}
\description{
ggmatrix
}
