% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ggthemeHiContacts}
\alias{ggthemeHiContacts}
\title{ggplot2-related functions}
\usage{
ggthemeHiContacts(ticks = TRUE)
}
\arguments{
\item{ticks}{ticks}
}
\value{
a custom ggplot2 theme
}
\description{
ggplot2-related functions
}
