% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggthemes.R
\name{ggthemeHiContacts}
\alias{ggthemeHiContacts}
\title{ggthemeHiContacts}
\usage{
ggthemeHiContacts(ticks = TRUE)
}
\arguments{
\item{ticks}{ticks}
}
\value{
a custom ggplot2 theme
}
\description{
ggthemeHiContacts
}
