% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggthemes.R
\name{ggtheme_HiContacts}
\alias{ggtheme_HiContacts}
\alias{ggtheme_HiContacts_tracks}
\title{ggtheme_HiContacts}
\usage{
ggtheme_HiContacts(ticks = TRUE)

ggtheme_HiContacts_tracks()
}
\arguments{
\item{ticks}{ticks}
}
\value{
a custom ggplot2 theme

a custom ggplot2 theme
}
\description{
ggtheme_HiContacts
}
