% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg.R
\name{ggtiltedmatrix}
\alias{ggtiltedmatrix}
\title{ggtiltedmatrix}
\usage{
ggtiltedmatrix(mat_, ticks = TRUE, cols = afmhotr_colors, limits)
}
\arguments{
\item{mat_}{mat_}

\item{ticks}{ticks}

\item{cols}{cols}
}
\description{
ggtiltedmatrix
}
