% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lsCoolResolutions}
\alias{lsCoolResolutions}
\title{lsCoolResolutions}
\usage{
lsCoolResolutions(file, full.list = FALSE)
}
\arguments{
\item{file}{file}

\item{full.list}{full.list}
}
\description{
lsCoolResolutions
}
