% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{pairs2gi}
\alias{pairs2gi}
\title{readPairs}
\usage{
pairs2gi(
  file,
  chr1.field = 2,
  start1.field = 3,
  chr2.field = 4,
  start2.field = 5,
  nThread = 16,
  nrows = Inf
)
}
\arguments{
\item{file}{pairs file: \verb{<readname>\\t<chr1>\\t<start1>\\t<chr2>\\t<start2>}}
}
\description{
readPairs
}
