% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{palettes}
\alias{palettes}
\alias{bwrColors}
\alias{bbrColors}
\alias{bgrColors}
\alias{afmhotrColors}
\alias{coolerColors}
\alias{rainbowColors}
\title{Matrix palettes}
\usage{
bwrColors()

bbrColors()

bgrColors()

afmhotrColors()

coolerColors()

rainbowColors()
}
\value{
A vector of colours carefully
picked for Hi-C contact heatmaps
}
\description{
Matrix palettes
}
\examples{
bwrColors()
bbrColors()
bgrColors()
afmhotrColors()
coolerColors()
rainbowColors()
}
