% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{bwrColors}
\alias{bwrColors}
\alias{afmhotrColors}
\alias{bbrColors}
\title{Matrix palettes}
\usage{
bwrColors()

afmhotrColors()

bbrColors()
}
\value{
ggplot
}
\description{
Matrix palettes
}
\examples{
bwrColors()
afmhotrColors()
bbrColors()
}
