% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{bwrColors}
\alias{bwrColors}
\alias{afmhotrColors}
\alias{coolerColors}
\alias{bbrColors}
\alias{bgrColors}
\title{Matrix palettes}
\usage{
bwrColors()

afmhotrColors()

coolerColors()

bbrColors()

bgrColors()
}
\value{
ggplot
}
\description{
Matrix palettes
}
\examples{
bwrColors()
afmhotrColors()
coolerColors()
bbrColors()
bgrColors()
}
