% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMatrix.R
\name{bwr_colors}
\alias{bwr_colors}
\alias{afmhotr_colors}
\alias{bbr_colors}
\title{Matrix palettes}
\usage{
bwr_colors()

afmhotr_colors()

bbr_colors()
}
\value{
ggplot
}
\description{
Matrix palettes
}
\examples{
bwr_colors()
afmhotr_colors()
bbr_colors()
}
