% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotMatrix}
\alias{plotMatrix}
\title{plotMatrix}
\usage{
plotMatrix(
  gis,
  colour_by = "score",
  scale = "log10",
  loops = NULL,
  borders = NULL,
  limits = NULL,
  dpi = 500,
  rasterize = TRUE,
  symmetrical = TRUE,
  chrom_lines = TRUE,
  cmap = NULL
)
}
\arguments{
\item{gis}{gis}

\item{scale}{scale}

\item{limits}{limits}

\item{dpi}{dpi}

\item{rasterize}{rasterize}

\item{symmetrical}{symmetrical}

\item{chrom_lines}{chrom_lines}

\item{cmap}{color map}
}
\description{
plotMatrix
}
