% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMatrix.R
\name{plotMatrix}
\alias{plotMatrix}
\title{plotMatrix}
\usage{
plotMatrix(
  x,
  use.assay = "balanced",
  scale = "log10",
  loops = NULL,
  borders = NULL,
  limits = NULL,
  dpi = 500,
  rasterize = TRUE,
  symmetrical = TRUE,
  chrom_lines = TRUE,
  cmap = NULL
)
}
\arguments{
\item{x}{x}

\item{use.assay}{use.assay}

\item{scale}{scale}

\item{loops}{loops}

\item{borders}{borders}

\item{limits}{limits}

\item{dpi}{dpi}

\item{rasterize}{rasterize}

\item{symmetrical}{symmetrical}

\item{chrom_lines}{chrom_lines}

\item{cmap}{color map}
}
\value{
ggplot
}
\description{
plotMatrix
}
\examples{
library(HiContacts)
data(contacts_yeast)
plotMatrix(
    contacts_yeast, 
    use.assay = 'balanced', 
    scale = 'log10', 
    limits = c(-4, -1)
)
}
