% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotMatrixList}
\alias{plotMatrixList}
\title{plotMatrixList}
\usage{
plotMatrixList(ls, truncate_tip = 0.2, ...)
}
\arguments{
\item{ls}{ls}

\item{truncate_tip}{truncate_tip}

\item{...}{...}
}
\description{
plotMatrixList
}
