% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotTriangularMatrix}
\alias{plotTriangularMatrix}
\title{plotTriangularMatrix}
\usage{
plotTriangularMatrix(gis, limits = NULL, dist_max = NULL)
}
\arguments{
\item{gis}{gis}

\item{limits}{limits}

\item{dist_max}{dist_max}
}
\description{
plotTriangularMatrix
}
