% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotTriangularMatrix}
\alias{plotTriangularMatrix}
\title{plotTriangularMatrix}
\usage{
plotTriangularMatrix(gis, limits = NULL, truncate_tip = 0.2)
}
\arguments{
\item{gis}{gis}

\item{limits}{limits}

\item{truncate_tip}{truncate_tip}
}
\description{
plotTriangularMatrix
}
