% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ps.R
\name{distanceLaw}
\alias{distanceLaw}
\alias{PsBreaks}
\title{Compute the law of distance-dependent contact frequency, a.k.a. P(s)}
\usage{
distanceLaw(
  x,
  by_chr = FALSE,
  filtered_chr = c("XII", "chrXII", "chr12", "12", "Mito", "MT", "chrM")
)

PsBreaks()
}
\arguments{
\item{x}{A \code{Contacts} object}

\item{by_chr}{by_chr}

\item{filtered_chr}{filtered_chr}
}
\value{
a tibble

tbl
}
\description{
P(s) will be approximated if no pairs are provided, or the exact P(s)
will be computed if a \code{.pairs} file is added to the \code{Contacts} object
using \code{pairsFile(x) <- "..."}.
}
\examples{
library(HiContacts)
contacts_yeast <- contacts_yeast()
ps <- distanceLaw(contacts_yeast)
ps
}
