% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ps.R
\name{distanceLaw}
\alias{distanceLaw}
\alias{localDistanceLaw}
\alias{PsBreaks}
\title{Compute the law of distance-dependent contact frequency, a.k.a. P(s)}
\usage{
distanceLaw(
  x,
  by_chr = FALSE,
  filtered_chr = c("XII", "chrXII", "chr12", "12", "Mito", "MT", "chrM")
)

localDistanceLaw(x, coords = coords)

PsBreaks()
}
\arguments{
\item{x}{A \code{HiCExperiment} object}

\item{by_chr}{by_chr}

\item{filtered_chr}{filtered_chr}

\item{coords}{GRanges}
}
\value{
a tibble

a tibble

tbl
}
\description{
P(s) will be approximated if no pairs are provided, or the exact P(s)
will be computed if a \code{.pairs} file is added to the \code{HiCExperiment} object
using \code{pairsFile(x) <- "..."}.
}
\examples{
library(HiContacts)
contacts_yeast <- contacts_yeast()
ps <- distanceLaw(contacts_yeast)
ps
local_ps <- localDistanceLaw(
    contacts_yeast,
    GenomicRanges::GRanges(
        c("telomere" = "II:1-20000", "arm" = "II:300000-700000")
    )
)
local_ps
}
