% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmetics.R
\name{smoothen}
\alias{smoothen}
\title{smoothen}
\usage{
smoothen(
  x,
  use.assay = "balanced",
  use_serpentine_trend = TRUE,
  serpentine_niter = 10L,
  serpentine_ncores = 16L
)
}
\arguments{
\item{x}{a \code{contacts} object}

\item{use.assay}{use.assay}

\item{use_serpentine_trend}{whether to use the trend estimated with
serpentine (this requires \code{reticulate} and the python package \code{serpentine})}

\item{serpentine_niter}{number of iterations to use for serpentine}

\item{serpentine_ncores}{number of CPUs to use for serpentine}
}
\description{
smoothen
}
