% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4C.R
\name{virtual4C}
\alias{virtual4C}
\title{virtual4C}
\usage{
virtual4C(x, viewpoint, use.assay = "balanced")
}
\arguments{
\item{x}{a \code{contacts} object}

\item{viewpoint}{viewpoint defined as a GRanges}

\item{use.assay}{use.assay}
}
\value{
A tibble with the contact frequency of the viewpoint, per bin
along the imported genomic range.
}
\description{
virtual4C
}
\examples{
library(HiContacts)
data(contacts_yeast)
virtual4C(contacts_yeast, GRanges('II:490000-510000'))
}
